refactor: ticket -> ticketing 폴더명 변경
