--- ticketing/src/main/java/org/truve/platform/ticketing/service/schedule/dto/AdmissionTokenClaimsDTO.java
@@ -1,16 +1,0 @@
-package org.truve.platform.ticketing.service.schedule.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class AdmissionTokenClaimsDTO {
-	private Long userId;
-	private Long showId;
-	private String tokenType;
-
-	public static AdmissionTokenClaimsDTO of(Long userId, Long showId, String tokenType) {
-		return new  AdmissionTokenClaimsDTO(userId, showId, tokenType);
-	}
-}
