--- ticketing/src/main/java/org/truve/platform/ticketing/service/schedule/dto/SeatSectionsDto.java
@@ -1,19 +1,0 @@
-package org.truve.platform.ticketing.service.schedule.dto;
-
-import org.truve.platform.ticketing.service.schedule.constant.SeatStatus;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class SeatSectionsDto {
-	Long sectionId;
-	String sectionName;
-	String grade;
-	Long price;
-	Long seatId;
-	String row;
-	Long col;
-	SeatStatus status;
-}
