--- ticketing/src/main/java/org/truve/platform/ticketing/service/schedule/repository/SeatSectionRepository.java
@@ -1,7 +1,0 @@
-package org.truve.platform.ticketing.service.schedule.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.truve.platform.ticketing.service.schedule.domain.entity.SeatSection;
-
-public interface SeatSectionRepository extends JpaRepository<SeatSection, Long> {
-}
