+++ ticketing/src/main/java/org/truve/platform/ticketing/service/ticketing/repository/ShowScheduledRepository.java
@@ -1,0 +1,7 @@
+package org.truve.platform.ticketing.service.ticketing.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.truve.platform.ticketing.service.ticketing.domain.entity.ShowScheduled;
+
+public interface ShowScheduledRepository extends JpaRepository<ShowScheduled, Long> {
+}
